reworked as described in docs (#3413)
