+++ examples/pytests/tests/conftest.py
@@ -1,5 +1,20 @@
-"""
-NOTE: This is uncommented because pytest doesn't allow for conftest.py files in subdirectories.
-      If you want to use this example as a template, you have to uncomment the next line
-"""
-# pytest_plugins = ['nicegui.testing.fixtures']
+from typing import Generator
+
+import pytest
+from pytests.startup import startup
+
+from nicegui.testing import Screen, User
+
+pytest_plugins = ['nicegui.testing.fixtures']
+
+
+@pytest.fixture
+def user(user: User) -> Generator[User, None, None]:
+    startup()
+    yield user
+
+
+@pytest.fixture
+def screen(screen: Screen) -> Generator[Screen, None, None]:
+    startup()
+    yield screen
